Make C/C++ code compile properly on arm 32bit.
Most of the warnings are about '%lu' referring to different sizes on

32bit vs 64bit platforms. Using the PRIu64 fixes this

(https://matt.sh/howto-c#_printing-types)



There's also a warning about comparing signed ints with unsigned ints,

which this fixes with a cast.

+++ endhost/ssp/SCIONProtocol.cpp
@@ -727,7 +727,7 @@ void SSPProtocol::handleTimerEvent()
     struct timeval current;
     gettimeofday(&current, NULL);
     mConnectionManager->handleTimeout();
-    if (elapsedTime(&mLastProbeTime, &current) >= mProbeInterval) {
+    if (elapsedTime(&mLastProbeTime, &current) >= (int32_t)mProbeInterval) {
         mConnectionManager->sendProbes(++mProbeNum, mFlowID);
         mLastProbeTime = current;
     }
